requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
gunicorn==20.0.4
numpy==1.18.3
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
psycopg2==2.7.2
